examples/freeinpaint/requirements.txt
torchvision
transformers==4.46.3
ftfy
tensorboard
datasets
Pillow==9.5.0 
opencv-python 
imgaug 
accelerate==0.20.3
image-reward
hpsv2
torchmetrics
open-clip-torch
openai-clip
gradio==3.50.0
segment_anything
huggingface_hub==0.23.2
munch==4.0.0
pycocotools
xformers==0.0.22